+++ nata-monkey/src/main/java/com/nata/utils/HttpUtil.java
@@ -9,6 +9,7 @@ import org.apache.http.impl.client.HttpClients;
 import org.apache.http.message.BasicNameValuePair;
 
 import java.io.IOException;
+import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -19,16 +20,37 @@ import java.util.List;
  */
 public class HttpUtil {
 
-    public static void post(String message) throws IOException {
+    public static void post(String url,ArrayList<String> messages) throws IOException {
         CloseableHttpClient client = HttpClients.createDefault();
-        HttpPost httpPost = new HttpPost("http://localhost:3000/");
+        HttpPost httpPost = new HttpPost(url);
 
         List<NameValuePair> params = new ArrayList<>();
-        params.add(new BasicNameValuePair("message", message));
+        for (String message: messages
+             ) {
+            params.add(new BasicNameValuePair("message", message));
+        }
+
         httpPost.setEntity(new UrlEncodedFormEntity(params));
 
         CloseableHttpResponse response = client.execute(httpPost);
         client.close();
+    }
+
+
+    public static void postTestResult(ArrayList<String> messages) throws IOException {
+        post("http://localhost:3000/test",messages);
+    }
+
+    public static void postLive(String message) throws IOException {
+        ArrayList<String> messages = new ArrayList<>();
+        messages.add(message);
+        post("http://localhost:3000/live",messages);
+    }
+
+    public static void postSummary(String summary) throws IOException {
+        ArrayList<String> messages = new ArrayList<>();
+        messages.add(summary);
+        post("http://localhost:3000/summary",messages);
     }
 
 }